refactor: modularize TypeSense implementation
- Move TypeSense functionality to external/typesense module\n- Create specialized modules for types, client, and operations\n- Maintain backward compatibility through re-exports\n- Improve type safety with better error handling
