+++ apps/egghead/src/lib/external/typesense/index.ts
@@ -1,2 +1,12 @@
-// Re-export everything from the TypeSense module
+/**
+ * TypeSense module for indexing and searching content
+ */
+
+// Re-export all the types
+export * from './types'
+
+// Re-export client utilities
+export * from './client'
+
+// Re-export post operations
 export * from './post'
